amount date and setting webview done

+++ app/src/main/java/com/example/baji/HomeActivity/MatchesFragment/MatchesProfile/BottomFragment/AcceptBaji/AcceptBajiBottomFragment.java
@@ -66,7 +66,7 @@ public class AcceptBajiBottomFragment extends BottomSheetDialogFragment {
         super.onViewCreated(view, savedInstanceState);
         ButterKnife.bind(this,view);
         match=(Match)getArguments().getSerializable("match");
-        bajiOfferByAmount=getArguments().getDouble("amount");
+        bajiOfferByAmount=getArguments().getInt("amount");
         openBajiId=getArguments().getInt("openBajiId");
         bajiOfferByTeamName=getArguments().getString("teamName");
         bajiOfferByUsername=getArguments().getString("username");
